app/requirements.txt
pandas==2.3.1
python-dotenv==1.1.1
kagglehub==0.3.12
confluent-kafka==2.11.0
